web | basic auth with jwt

+++ Semestre_4_2022/Programação_Web/Aula_12_API/docker-lamp-pw/www/api/index.php
@@ -11,23 +11,38 @@ $uriSegments = explode("/", $rout);
 if(isset($uriSegments[1])){
 	switch($uriSegments[1]){
 		case 'client':
-			require_once("Controllers/ClientController.php");
-			$clientController = new ClientController();
+			require_once("Controllers/UserController.php");
+			$user = new UserController();
+			if($user -> isAdmin()){
+				require_once("Controllers/ClientController.php");
+				$clientController = new ClientController();
+				switch($request_method){
+					case 'GET':
+						if(!isset($uriSegments[2]))
+							$clientController -> listClients();
+						else
+							$clientController -> clientDetails($uriSegments[2]);
+					break;
+					case 'POST':
+						$clientController -> insertClient();
+					break;
+					case 'PUT':
+						$clientController -> updateClient($uriSegments[2]);
+					break;
+					case 'DELETE':
+						$clientController -> deleteClient($uriSegments[2]);
+					break;
+				}
+			}
+		break;
+
+
+		case 'user':
+			require_once('Controllers/UserController.php');
+			$user = new UserController();
 			switch($request_method){
 				case 'GET':
-					if(!isset($uriSegments[2]))
-						$clientController -> listClients();
-					else
-						$clientController -> clientDetails($uriSegments[2]);
-				break;
-				case 'POST':
-					$clientController -> insertClient();
-				break;
-				case 'PUT':
-					$clientController -> updateClient($uriSegments[2]);
-				break;
-				case 'DELETE':
-					$clientController -> deleteClient($uriSegments[2]);
+					$user -> validateLogin();
 				break;
 			}
 		break;
